+++ CHANGELOG.md
@@ -14,11 +14,12 @@
 
 * The following components have been refactored to meet best practice standards and pass linting.
 * If you have overridden any internal methods of these components, you may need to update your code.
+  - Action Toolbar
+  - Alert
+  - Animated Menu Button
+  - App-Wrapper
   - Button
   - Create
-  - Alert
-  - App-Wrapper
-  - Action Toolbar
 
 ## Component Enhancements